% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mtrl}
\alias{mtrl}
\title{物料处理的主表}
\usage{
mtrl(file = system.file("extdata/tpl.xlsx", package = "kdcmtrlpkg"),
  keepInput = FALSE)
}
\arguments{
\item{file}{输入的模板文件}

\item{keepInput}{是否保留输入数据否认为否}
}
\value{
返回list
}
\description{
物料处理的主表
}
\examples{
mtrl();
}
