% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_genDataCalc}
\alias{mtrl_genDataCalc}
\title{用于处理内部字段}
\usage{
mtrl_genDataCalc(metaInfo, inputData)
}
\arguments{
\item{metaInfo}{元数据}

\item{inputData}{输入数据}
}
\value{
返回值
}
\description{
用于处理内部字段
}
\examples{
mtrl_genDataCalc();
}
