% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_genDataCopy}
\alias{mtrl_genDataCopy}
\title{获取组织的复制信息}
\usage{
mtrl_genDataCopy(metaInfo, inputData)
}
\arguments{
\item{metaInfo}{模板元数据}

\item{inputData}{物料填写表}
}
\value{
返回值
}
\description{
获取组织的复制信息
}
\examples{
mtrl_genDataCopy();
}
