% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_genDataInput}
\alias{mtrl_genDataInput}
\title{针对数据进行命名处理}
\usage{
mtrl_genDataInput(metaInfo, inputData)
}
\arguments{
\item{metaInfo}{元数据}

\item{inputData}{输入数据}
}
\value{
返回值
}
\description{
针对数据进行命名处理
}
\examples{
mtrl_genDataInput();
}
