% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_genDataKdc}
\alias{mtrl_genDataKdc}
\title{根据配置表产生金蝶云固有字段并按行复制}
\usage{
mtrl_genDataKdc(metaInfo, inputData)
}
\arguments{
\item{metaInfo}{全部元数据表}

\item{inputData}{输入数据}
}
\value{
返回值
}
\description{
根据配置表产生金蝶云固有字段并按行复制
}
\examples{
mtrl_genDataKdc();
}
