% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_genDataRef}
\alias{mtrl_genDataRef}
\title{获取物料中的复制数据}
\usage{
mtrl_genDataRef(metaInfo, inputData)
}
\arguments{
\item{metaInfo}{元数据}

\item{inputData}{输入数据}
}
\value{
返回参数数据的DF
}
\description{
获取物料中的复制数据
}
\examples{
mtrl_genDataRef();
}
