% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_getInputData}
\alias{mtrl_getInputData}
\title{获取数据填写数据表实际数据}
\usage{
mtrl_getInputData(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{数据填写表}
}
\value{
返回值
}
\description{
获取数据填写数据表实际数据
}
\examples{
mtrl_getInputData();
}
