% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_getMetaInfo}
\alias{mtrl_getMetaInfo}
\title{获取规则表信息}
\usage{
mtrl_getMetaInfo(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{物料模板所在的目前}
}
\value{
返回列表
}
\description{
获取规则表信息
}
\examples{
mtrl_getMetaInfo();
system.file('extdata/tpl.xlsx',package = 'kdcmtrlpkg');
}
