% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaSetTitleInfo.R
\name{mtrl_getOrgInfo}
\alias{mtrl_getOrgInfo}
\title{获取物料模板的组织信息}
\usage{
mtrl_getOrgInfo(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{物料模板所有的路径}
}
\value{
返回包含组织信息的list
}
\description{
获取物料模板的组织信息
}
\examples{
mtrl_getOrgInfo()
}
