% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaSetTitleInfo.R
\name{mtrl_getSheetNames}
\alias{mtrl_getSheetNames}
\title{读取物料导入模板各个业签的名称}
\usage{
mtrl_getSheetNames(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{物料模板所在的文件路径}
}
\value{
返回页答的list
}
\description{
读取物料导入模板各个业签的名称
}
\examples{
mtrl_getSheetNames();
}
