% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbReadRuleTable.R
\name{mtrl_getTplName}
\alias{mtrl_getTplName}
\title{获取数据的输出模板字段}
\usage{
mtrl_getTplName(metaInfo)
}
\arguments{
\item{metaInfo}{元数据}
}
\value{
返回值
}
\description{
获取数据的输出模板字段
}
\examples{
mtrl_getTplName();
}
