% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mtrl_main}
\alias{mtrl_main}
\title{处理物料主表数据}
\usage{
mtrl_main(file = system.file("extdata/tpl.xlsx", package = "kdcmtrlpkg"))
}
\arguments{
\item{file}{文件所在的路径}
}
\value{
返回值
}
\description{
处理物料主表数据
}
\examples{
mtrl_main()
}
