% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccReadFixedTable.R
\name{mtrl_sheet1_readme}
\alias{mtrl_sheet1_readme}
\title{读取物料单页签的导入说明}
\usage{
mtrl_sheet1_readme(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{文件所有的路径}
}
\value{
返回值
}
\description{
读取物料单页签的导入说明
}
\examples{
mtrl_sheet1_readme()
}
