% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccReadFixedTable.R
\name{mtrl_sheet2_mtrlgp}
\alias{mtrl_sheet2_mtrlgp}
\title{读取物料分组数据}
\usage{
mtrl_sheet2_mtrlgp(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{物料导入模板所有路径}
}
\value{
返回值
}
\description{
读取物料分组数据
}
\examples{
mtrl_sheet2_mtrlgp()
}
