% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccReadFixedTable.R
\name{mtrl_sheet4_ruleTable}
\alias{mtrl_sheet4_ruleTable}
\title{读取规则表}
\usage{
mtrl_sheet4_ruleTable(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{物料导入模板所有路径}
}
\value{
返回值
}
\description{
读取规则表
}
\examples{
mtrl_sheet4_ruleTable()
}
