% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccReadFixedTable.R
\name{mtrl_sheet5_setHeader}
\alias{mtrl_sheet5_setHeader}
\title{读取设置表头}
\usage{
mtrl_sheet5_setHeader(file = system.file("extdata/tpl.xlsx", package =
  "kdcmtrlpkg"))
}
\arguments{
\item{file}{物料导入模板所有路径}
}
\value{
返回值
}
\description{
读取设置表头
}
\examples{
mtrl_sheet5_setHeader()
}
